Use ReaderT for ColumnFormatter
Why?
====

When printing results, the column formatter has to be configured at the
topmost level (where it has all result data) to calculate widths
appropriately; however, it's only used layers deep, when rendering the
columns themselves.

This moves the formatter into a ReaderT so the configuration can be
passed around appropriately.