Add Readable typeclass to simplify safe reading of files
Why?
====

After the introduction of:

    safeReadFile :: FilePath -> IO (Either E.IOException String)

There were places who needed to repack data into the appropriate type.
This is an extra step and, while in these situations not necessarily
*slow*, should be removed.

This introduces a Readable typeclass which dictates implementation of
`readFile'` based on the resulting requested type. This means `safeReadFile`
can be used in various situations without having to `pack`.