Improve alias calculations
Previously, aliases were calculated twice:

1. when building out the list of terms to search for
2. when merging results into TermResults for grouping (and likelihood calcs)

This problem was masked by the fact that wildcard matching on both ends
was the same, with '%s'. However, the goal is to move away from simple
'%s' matching and do a system where transformations can be applied to
the text, which would address #18.

This moves us one step closer to fixing that issue.